Create README.md
Python program uses os.walk recursion to process a specified Unix directory and any/all subordinate or nested Unix directories. Every non-directory file that is encountered is parsed as a sequence of words. A list of words that are recognized as palindromes is constructed. The output is a list of all of the palindromes that were found and how many occurrences of each palindrome.